Terminado el ejemplo de queue

+++ main/src/queue/queue.java
@@ -2,10 +2,10 @@ package queue;
 
 public interface queue {
 
-    void enqueue(double data);
+    void enqueue(int data);
     void dequeue();
     void queueDisplay();
-    void queueSearch(double search);
+    void queueSearch(int search);
     void queueEmpty();
     void queueAverage();
     void queueMax();
